requirements.txt
streamlit
pandas
scikit-learn
spacy
nltk
sentence-transformers
pdfminer.six
